Decadac in Custom-Instr
Making the teh decadac driver more streamlined so we only need to change things in the config file.

The code is not pretty with 'exec' but dont know a better way to code it - Just want to have all settings in the config file rather than both places.

Note new format in the instr.config file.